requirements.txt
chrome_version==0.3.1
pytest==8.3.4
selenium==4.29.0
webdriver_manager==4.0.2

allure-python-commons~=2.13.5
webdriver-manager~=4.0.2